# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EVO Platform API",
    "version": "3.0.0",
    "description": "EVO Platform by NuevaCore - Multi-Cloud Security & Cost Management Platform. Supports AWS and Azure."
  },
  "servers": [
    {
      "url": "https://api.evo.nuevacore.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/api/functions/cost-optimization": {
      "post": {
        "summary": "Get cost optimization recommendations",
        "operationId": "cost_optimization",
        "tags": [
          "Cost"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Operation success status"
                    },
                    "data": {
                      "type": "object",
                      "description": "Response data"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid input parameters"
          },
          "401": {
            "description": "Unauthorized - Missing or invalid authentication"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Not found - Resource does not exist"
          },
          "429": {
            "description": "Too many requests - Rate limit exceeded"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/functions/azure-cost-optimization": {
      "post": {
        "summary": "Azure Advisor cost recommendations",
        "operationId": "azure_cost_optimization",
        "tags": [
          "Azure",
          "Cost"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Operation success status"
                    },
                    "data": {
                      "type": "object",
                      "description": "Response data"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid input parameters"
          },
          "401": {
            "description": "Unauthorized - Missing or invalid authentication"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Not found - Resource does not exist"
          },
          "429": {
            "description": "Too many requests - Rate limit exceeded"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "AWS Cognito JWT token from Authorization header"
      }
    }
  },
...
}